add control button - set auto - report

+++ bin/main/static/js/detailroom.js
@@ -40,18 +40,21 @@ function createTableDevice(datasource){
                 "delete": {
                     enabled: true,
                     template: function (item) {
-                        $.ajax({
-                            async : false,
-                            method: "delete",
-                            data: JSON.stringify({id: item.id, ip: item.ip, nameDevice: item.nameDevice, state: item.state, type: item.type}),
-                            contentType: "application/json",
-                            url: "http://localhost/smarthome/deletedevice/"+$(".detailroomname").text()
-                        }).done(function(data, textStatus, xhr){
+                        var confirmDelete = confirm("Are you sure deletr device with IP = " + item.ip )
+                        if(confirmDelete === true){
+                            $.ajax({
+                                async : false,
+                                method: "delete",
+                                // data: JSON.stringify({id: item.id}),
+                                contentType: "application/json",
+                                url: "http://localhost/smarthome/deletedevice/"+item.id
+                            }).done(function(data, textStatus, xhr){
                                 status_create = xhr.status;
-                        }).fail(function(data, textStatus, xhr){
+                            }).fail(function(data, textStatus, xhr){
                                 status_create = data.status;
-                        });
-                        return "Delete row with ID = " + item.id
+                            });
+                            return "Delete row with ID = " + item.id;  
+                        }
                     }
                 }
             }
